Refactor for new JSON writer
